delete connections whose airbyte status is "deprecated"
celery task to delete a list of connections

cleanup: new dbthelpers.py module with create_or_update_org_cli_block from airbytehelpers
