+++ ddpui/tests/core/test_celery_tasks.py
@@ -358,10 +358,15 @@ def test_detect_schema_changes_for_org_ensure_orphan_connections_are_deleted(
     assert OrgSchemaChange.objects.filter(org=org_without_workspace).count() == 1
     with patch(
         "ddpui.ddpairbyte.airbytehelpers.fetch_and_update_org_schema_changes"
-    ) as fetch_and_update_org_schema_changes_mock:
+    ) as fetch_and_update_org_schema_changes_mock, patch(
+        "ddpui.celeryworkers.tasks.send_text_message"
+    ) as mock_send_text_message:
         fetch_and_update_org_schema_changes_mock.return_value = None, "error"
         detect_schema_changes_for_org(org_without_workspace)
     assert OrgSchemaChange.objects.filter(org=org_without_workspace).count() == 0
+    mock_send_text_message.assert_called_once_with(
+        "adminemail", "Schema change detection errors for test-org-WO-slug", "error"
+    )
 
 
 def test_get_connection_catalog_task_error(org_without_workspace: Org):
@@ -369,7 +374,9 @@ def test_get_connection_catalog_task_error(org_without_workspace: Org):
     task_key = "test-task-key"
     with patch(
         "ddpui.ddpairbyte.airbytehelpers.fetch_and_update_org_schema_changes"
-    ) as fetch_and_update_org_schema_changes_mock:
+    ) as fetch_and_update_org_schema_changes_mock, patch(
+        "ddpui.celeryworkers.tasks.send_text_message"
+    ) as mock_send_text_message:
         fetch_and_update_org_schema_changes_mock.return_value = None, "error"
         get_connection_catalog_task(task_key, org_without_workspace.id, "fake-connection-id")
     result = SingleTaskProgress.fetch(task_key)
@@ -381,6 +388,11 @@ def test_get_connection_catalog_task_error(org_without_workspace: Org):
             "result": None,
         },
     ]
+    mock_send_text_message.assert_called_once_with(
+        os.getenv("ADMIN_EMAIL"),
+        f"Unhandled schema change detection errors for {org_without_workspace.slug}",
+        "error",
+    )
 
 
 def test_get_connection_catalog_task_success(org_without_workspace: Org):
